Opening the window wider
The bad database state is allowed by having the "is this user in
a cohort" check and the "add this user to a cohort" action not be
in an atomic block.

By adding this sleep statement, it becomes easier to exploit the
timing window and add a previously un-cohorted earner to mutiple
cohorts.

Confirmed on local devstack by uploading a csv to assign a user to
cohort group 1, then manually adding them to group 2 during the
window. User ends up in both cohorts, which should not be possible.